refactor: Upgrade command-line scripts to use ESM-style
