+++ package.json
@@ -5,11 +5,11 @@
     "prepare": "husky",
     "setup": "cp -i .env.example .env && npm run prepare",
     "start": "ng serve",
-    "set-env-variables": "node ./scripts/set-env-variables.js",
+    "set-env-variables": "node ./scripts/set-env-variables.mjs",
     "prebuild": "npm run set-env-variables",
     "build": "ng build",
     "build:production": "npm run build -- --configuration=production",
-    "postbuild:production": "node ./scripts/fix-output-path.js",
+    "postbuild:production": "node ./scripts/fix-output-path.mjs",
     "build:e2e": "npm run build:production -- --configuration=test",
     "preserve:production": "cp www/index.html www/404.html",
     "serve:production": "npx http-server ./www/ -c-1",
