+++ scripts/README.md
@@ -1,0 +1,5 @@
+# Command-line Scripts
+
+This directory contains various command-line scripts that can be used before/after building the application.
+
+See within each script for details on how and when to use.
