+++ src/environments/environment.prod.ts.template.mjs
@@ -1,5 +1,5 @@
 // Output should be a valid TS-file:
-module.exports = `// THIS FILE IS GENERATED BY 'npm run set-env-variables'
+export default `// THIS FILE IS GENERATED BY 'npm run set-env-variables'
 
 export const environment = {
   production: ${process.env.NG_PRODUCTION === 'true' || false},
